requirements.txt
numpy
pandas
librosa
matplotlib
scikit-learn
tensorflow
torch
torchaudio
jupyter
streamlit
requests
openai
python-dotenv
streamlit-webrtc
opencv-python-headless
av
